refactor(ws-do): split user auth and presence domains
