+++ src/ws-do/src/ws-durable.test.ts
@@ -1,1132 +1,59 @@
-import { describe, it, expect, vi, beforeEach } from "vitest"
-import { createMockCtx, createMockWebSocket } from "./__mocks__/cf"
-
-// --- Cloudflare Workers globals that don't exist in Node ---
-
-// Replace the global Response with one that allows status 101 and a webSocket property
-class CFResponse {
-  status: number
-  webSocket: unknown
-  private _body: BodyInit | null
-  private _headers: Headers
-
-  constructor(body: BodyInit | null = null, init: ResponseInit & { webSocket?: unknown } = {}) {
-    this._body = body
-    this._headers = new Headers(init.headers)
-    this.status = init.status ?? 200
-    this.webSocket = (init as { webSocket?: unknown }).webSocket
-  }
-
-  async text(): Promise<string> {
-    if (this._body == null) return ""
-    if (typeof this._body === "string") return this._body
-    return ""
-  }
-
-  async json(): Promise<unknown> {
-    return JSON.parse(await this.text())
-  }
-
-  get headers() { return this._headers }
-}
-
-globalThis.Response = CFResponse as unknown as typeof Response
-
-// WebSocketPair — creates a paired (client, server) mock
-globalThis.WebSocketPair = class {
-  0: ReturnType<typeof createMockWebSocket>
-  1: ReturnType<typeof createMockWebSocket>
-  constructor() {
-    this[0] = createMockWebSocket()
-    this[1] = createMockWebSocket()
-  }
-} as unknown as typeof WebSocketPair
-
-// WebSocketRequestResponsePair — used for the ping/pong auto-response
-globalThis.WebSocketRequestResponsePair = class {
-  constructor(public request: string, public response: string) { }
-} as unknown as typeof WebSocketRequestResponsePair
-
-// --- Module mocks ---
-
-// Mock cloudflare:workers DurableObject base class
-vi.mock("cloudflare:workers", () => ({
-  DurableObject: class {
-    ctx: unknown
-    env: unknown
-    constructor(ctx: unknown, env: unknown) {
-      this.ctx = ctx
-      this.env = env
-    }
-  },
-}))
-
-// Mock @alook/shared
-const mockGetValidSession = vi.fn<(db: unknown, token: string) => Promise<string | null>>()
-const mockGetValidSessionWithIdentity = vi.fn<(db: unknown, token: string) => Promise<{ userId: string; name: string; discriminator: string } | null>>()
-const mockGetMachineTokenByToken = vi.fn()
-const mockGetLatestTokenForUser = vi.fn()
-const mockGetRuntimeIdsByDaemon = vi.fn()
-const mockGetMachineByDaemon = vi.fn().mockResolvedValue({ daemonId: "my-daemon", workspaceId: "sp_ws1" })
-const mockCreateDb = vi.fn().mockReturnValue({})
-const mockHashCredential = vi.fn(async (bearer: string) => `hash:${bearer}`)
-const mockFindCredentialByHash = vi.fn()
-const mockGetMachineByIdForUser = vi.fn()
-const mockUpsertMachineByMachineId = vi.fn()
-const mockTouchMachineHeartbeat = vi.fn()
-const mockMarkMachineOffline = vi.fn()
-const mockMarkMachineOnlineIfOffline = vi.fn()
-const mockGetCoMemberUserIds = vi.fn<(db: unknown, userId: string) => Promise<string[]>>().mockResolvedValue([])
-const mockGetFriendUserIds = vi.fn<(db: unknown, userId: string) => Promise<string[]>>().mockResolvedValue([])
-const mockGetChannelForMember = vi.fn()
-const mockListChannelMemberUserIds = vi.fn<(db: unknown, channelId: string) => Promise<string[]>>().mockResolvedValue([])
-const mockIsChannelPrivate = vi.fn<(db: unknown, channelId: string) => Promise<boolean>>().mockResolvedValue(false)
-const mockGetPrivateChannelAudienceUserIds = vi.fn<(db: unknown, channelId: string) => Promise<string[]>>().mockResolvedValue([])
-const mockResolveScopeMemberUserIds = vi.fn<(db: unknown, opts: { scope: string; scopeId: string }) => Promise<string[]>>().mockResolvedValue([])
-// Default: non-thread channel → typing uses the shared resolver path.
-const mockGetChannelType = vi.fn<(db: unknown, channelId: string) => Promise<string | null>>().mockResolvedValue("text")
-const mockListThreadParticipantUserIds = vi.fn<(db: unknown, channelId: string) => Promise<string[]>>().mockResolvedValue([])
-const mockGetDM = vi.fn()
-const mockListMembers = vi.fn()
-const mockListBotsForMachine = vi.fn<(db: unknown, machineId: string) => Promise<Array<{ id: string; name: string; discriminator: string; description: string }>>>().mockResolvedValue([])
-const mockIsBotOnline = vi.fn<(db: unknown, botUserId: string) => Promise<boolean>>().mockResolvedValue(false)
-const mockGetBotBinding = vi.fn<(db: unknown, botId: string) => Promise<{ machineId: string; runtime: string } | null>>().mockResolvedValue(null)
-const mockGetBotBindingWithOwner = vi.fn<(db: unknown, botId: string) => Promise<{ machineId: string; runtime: string; ownerUserId: string; name: string; discriminator: string } | null>>().mockResolvedValue(null)
-const mockInsertBotActivityEventAndPrune = vi.fn<(db: unknown, data: unknown) => Promise<{ id: string; createdAt: string } | null>>().mockResolvedValue(null)
-const mockInsertBotAuditSessionReset = vi.fn<(db: unknown, data: unknown) => Promise<{ id: string; createdAt: string } | null>>().mockResolvedValue(null)
-const mockInsertBotAuditNap = vi.fn<(db: unknown, data: unknown) => Promise<{ id: string; createdAt: string } | null>>().mockResolvedValue(null)
-const mockInsertBotAuditModelChanged = vi.fn<(db: unknown, data: unknown) => Promise<{ id: string; createdAt: string } | null>>().mockResolvedValue(null)
-const mockInsertBotAuditProviderChanged = vi.fn<(db: unknown, data: unknown) => Promise<{ id: string; createdAt: string } | null>>().mockResolvedValue(null)
-const mockTouchBotRefreshContext = vi.fn<(db: unknown, botId: string, now: string) => Promise<void>>().mockResolvedValue(undefined)
-const mockUpdateProfile = vi
-  .fn<(db: unknown, userId: string, data: { statusEmoji?: string | null; statusText?: string | null }) => Promise<unknown>>()
-  .mockResolvedValue({})
-const mockGetProfile = vi
-  .fn<(db: unknown, userId: string) => Promise<{ statusEmoji: string | null; statusText: string | null } | null>>()
-  .mockResolvedValue(null)
-const mockReconcileBotActivityFromRunningAgents = vi
-  .fn<(db: unknown, machineId: string, runningAgentIds: string[]) => Promise<Array<{ botUserId: string; statusEmoji: string; statusText: string }>>>()
-  .mockResolvedValue([])
-const mockGetUserInternal = vi.fn<(db: unknown, id: string) => Promise<{ isBot: boolean; ownerUserId: string | null } | null>>().mockResolvedValue(null)
-// mockToSummary now returns row.status verbatim — status is the source of
-// truth on the column, not a derivation from lastSeenAt. See
-// plans/community-machine-presence-fix.md.
-const mockToSummary = vi.fn((row: any) => ({
-  id: row.id,
-  hostname: row.hostname ?? "",
-  displayName: row.displayName ?? row.hostname ?? "",
-  platform: row.platform ?? "",
-  arch: row.arch ?? "",
-  osRelease: row.osRelease ?? "",
-  daemonVersion: row.daemonVersion ?? "",
-  lastSeenAt: row.lastSeenAt ?? null,
-  status: (row.status as "online" | "offline") ?? "offline",
-  availableRuntimes: row.availableRuntimes ?? [],
-  createdAt: row.createdAt ?? "",
-  updatedAt: row.updatedAt ?? "",
-}))
-
-vi.mock("@alook/shared", () => {
-  const noopLogger = {
-    debug: () => { },
-    info: () => { },
-    warn: () => { },
-    error: () => { },
-    child: () => noopLogger,
-  }
-  // Bare-minimum safeParse stubs — the DO only calls `.safeParse(msg)` and
-  // reads `.success` / `.data`. Enough to route the test frames correctly
-  // without pulling in zod (which isn't a direct dep of @alook/ws-do).
-  const SessionErrorFrameSchema = {
-    safeParse(v: unknown) {
-      const m = v as { type?: unknown; code?: unknown; agentId?: unknown; launchId?: unknown; payload?: unknown }
-      if (m?.type !== "session.error" || m?.code !== "runtime_not_available") {
-        return { success: false } as const
-      }
-      return {
-        success: true as const,
-        data: {
-          type: "session.error" as const,
-          code: "runtime_not_available" as const,
-          agentId: typeof m.agentId === "string" ? (m.agentId as string) : undefined,
-          // launchId rides the frame now (Melisa's source half) — the DO uses
-          // it to evict pending reset/nap attribution on cold-start branch A.
-          launchId: typeof m.launchId === "string" ? (m.launchId as string) : undefined,
-          payload: (m.payload as Record<string, unknown> | undefined) ?? undefined,
-        },
-      }
-    },
-  }
-  // Mirror the shared HostReadyMessageSchema: the daemon's ready frame must be
-  // FLAT (fields at top level), not wrapped in a `ready` key. A wrapped frame
-  // is rejected — regression guard against the wire-shape mismatch we hit
-  // when the daemon sent `{type:"ready", ready:{...}}` while the DO expected
-  // flat top-level fields.
-  const HostReadyMessageSchema = {
-    safeParse(v: unknown) {
-      const m = v as { type?: unknown; runtimeReport?: unknown; runningAgents?: unknown }
-      if (m?.type !== "ready") return { success: false } as const
-      if (!Array.isArray(m?.runtimeReport)) return { success: false } as const
-      const data: Record<string, unknown> = {
-        type: "ready",
-        runtimeReport: m.runtimeReport,
-        runningAgents: Array.isArray(m.runningAgents) ? m.runningAgents : [],
-      }
-      for (const k of ["hostname", "platform", "arch", "osRelease", "daemonVersion"]) {
-        const val = (m as Record<string, unknown>)[k]
-        if (typeof val === "string") data[k] = val
-      }
-      return { success: true as const, data }
-    },
-  }
-  const AgentActivityMessageSchema = {
-    safeParse(v: unknown) {
-      const m = v as { type?: unknown; agentId?: unknown; state?: unknown }
-      if (m?.type !== "agent_activity") return { success: false } as const
-      if (typeof m.agentId !== "string") return { success: false } as const
-      if (!["idle", "starting", "running", "stopping"].includes(m.state as string)) return { success: false } as const
-      return { success: true as const, data: { type: "agent_activity" as const, agentId: m.agentId, state: m.state } }
-    },
-  }
-  const AgentTypingMessageSchema = {
-    safeParse(v: unknown) {
-      const m = v as { type?: unknown; agentId?: unknown; channelId?: unknown }
-      if (m?.type !== "agent_typing") return { success: false } as const
-      if (typeof m.agentId !== "string" || m.agentId.length === 0) return { success: false } as const
-      if (typeof m.channelId !== "string" || m.channelId.length === 0) return { success: false } as const
-      return {
-        success: true as const,
-        data: { type: "agent_typing" as const, agentId: m.agentId, channelId: m.channelId },
-      }
-    },
-  }
-  const AgentTypingStopMessageSchema = {
-    safeParse(v: unknown) {
-      const m = v as { type?: unknown; agentId?: unknown; channelId?: unknown }
-      if (m?.type !== "agent_typing_stop") return { success: false } as const
-      if (typeof m.agentId !== "string" || m.agentId.length === 0) return { success: false } as const
-      if (typeof m.channelId !== "string" || m.channelId.length === 0) return { success: false } as const
-      return {
-        success: true as const,
-        data: { type: "agent_typing_stop" as const, agentId: m.agentId, channelId: m.channelId },
-      }
-    },
-  }
-  const AgentSessionMessageSchema = {
-    safeParse(v: unknown) {
-      const m = v as { type?: unknown; agentId?: unknown; sessionId?: unknown; launchId?: unknown }
-      if (m?.type !== "agent_session") return { success: false } as const
-      if (typeof m.agentId !== "string" || m.agentId.length === 0) return { success: false } as const
-      if (typeof m.sessionId !== "string" || m.sessionId.length === 0) return { success: false } as const
-      if (typeof m.launchId !== "string" || m.launchId.length === 0) return { success: false } as const
-      return {
-        success: true as const,
-        data: {
-          type: "agent_session" as const,
-          agentId: m.agentId,
-          sessionId: m.sessionId,
-          launchId: m.launchId,
-        },
-      }
-    },
-  }
-  const HostBotAuditEventFrameSchema = {
-    safeParse(v: unknown) {
-      const m = v as { type?: unknown; agentId?: unknown; sessionId?: unknown; launchId?: unknown; event?: unknown }
-      if (m?.type !== "bot_audit_event") return { success: false } as const
-      if (typeof m.agentId !== "string" || m.agentId.length === 0) return { success: false } as const
-      const ev = m.event as { kind?: unknown; payload?: unknown }
-      if (!ev || typeof ev !== "object") return { success: false } as const
-      const kind = ev.kind
-      const payload = ev.payload as Record<string, unknown> | undefined
-      if (!payload || typeof payload !== "object") return { success: false } as const
-      let ok = false
-      if (kind === "cli_invocation") ok = typeof payload.subcommand === "string"
-      else if (kind === "tool_call") ok = typeof payload.name === "string"
-      else if (kind === "thinking")
-        ok = typeof payload.text === "string" && typeof payload.truncated === "boolean" && typeof payload.chars === "number"
-      if (!ok) return { success: false } as const
-      return {
-        success: true as const,
-        data: {
-          type: "bot_audit_event" as const,
-          agentId: m.agentId,
-          sessionId: typeof m.sessionId === "string" ? m.sessionId : m.sessionId === null ? null : undefined,
-          launchId: typeof m.launchId === "string" ? m.launchId : m.launchId === null ? null : undefined,
-          event: { kind, payload },
-        },
-      }
-    },
-  }
-  return {
-    // Real WS event-type strings the DO reads at runtime (#5 T2 — ws-do
-    // broadcasts now use WS_EVENTS.* instead of raw literals). Values match
-    // @alook/shared so the tests' `type: "community:*"` assertions still hold.
-    WS_EVENTS: {
-      TYPING_START: "community:typing.start",
-      TYPING_STOP: "community:typing.stop",
-      MACHINE_STATUS: "community:machine.status",
-      MACHINE_UPDATED: "community:machine.updated",
-      STATUS_UPDATE: "community:status.update",
-      BOT_AUDIT_EVENT: "community:bot.audit_event",
-      PRESENCE_UPDATE: "community:presence.update",
-    },
-    createDb: (d1: unknown) => mockCreateDb(d1),
-    createLogger: () => noopLogger,
-    withD1Retry: async <T>(fn: () => Promise<T>): Promise<T> => fn(),
-    // Real reach-trait impls — fanOutTyping's recipient split dispatches on
-    // channelReach (B2 reach axis single source) to route a thread to the
-    // participant set, a dm to its members, else the access audience. Mirror
-    // the real CHANNEL_TRAITS reach values + the stored-type guard.
-    channelReach: (t: string) =>
-      t === "thread"
-        ? "participant-set"
-        : t === "dm"
-          ? "dm-pair"
-          : "server-or-roster",
-    isStoredChannelType: (t: unknown) =>
-      t === "text" || t === "forum" || t === "thread" || t === "dm",
-    // Minimal `readOrStale` shim — bypasses the classifier so tests can
-    // inject arbitrary Error shapes at the query-fn boundary and observe
-    // fail-closed semantics. Real production behavior (retry then fallback)
-    // is covered by the shared `resilience.test.ts` suite.
-    readOrStale: async <T>(
-      fn: () => Promise<T>,
-      fallback: T,
-      _opts?: unknown,
-    ): Promise<{ value: T; stale: boolean }> => {
-      try { return { value: await fn(), stale: false } }
-      catch { return { value: fallback, stale: true } }
-    },
-    COMMUNITY_MACHINE_HEARTBEAT_MS: 60_000,
-    COMMUNITY_MACHINE_OFFLINE_THRESHOLD_MS: 120_000,
-    SessionErrorFrameSchema,
-    HostReadyMessageSchema,
-    AgentActivityMessageSchema,
-    AgentTypingMessageSchema,
-    AgentTypingStopMessageSchema,
-    AgentSessionMessageSchema,
-    HostBotAuditEventFrameSchema,
-    // Deterministic preset picker so the assertion can pin exact
-    // `statusEmoji`/`statusText` values regardless of the injected `seed`.
-    pickBotActivityPreset: (state: string) => {
-      if (state === "running") return { emoji: "⚡", text: "Working on it" }
-      if (state === "starting") return { emoji: "🌀", text: "Waking up" }
-      if (state === "stopping") return { emoji: "🌙", text: "Wrapping up" }
-      return { emoji: "💤", text: "Idle" }
-    },
-    RUNNING_PRESETS: [
-      { emoji: "⚡", text: "Working on it" },
-      { emoji: "🛠️", text: "Cooking" },
-      { emoji: "🧠", text: "Thinking hard" },
-      { emoji: "🔧", text: "Tinkering" },
-      { emoji: "🚀", text: "On it" },
-      { emoji: "🔥", text: "In the zone" },
-    ],
-    // Real impl of the write-path/reconciler guard — mirrors the shared module
-    // over the presets emitted above (idle/starting/stopping + running pool).
-    isBotActivityStatus: (emoji: string | null, text: string | null) => {
-      if (emoji === null && text === null) return false
-      const pairs = [
-        { emoji: "💤", text: "Idle" },
-        { emoji: "🌀", text: "Waking up" },
-        { emoji: "🌙", text: "Wrapping up" },
-        { emoji: "⚡", text: "Working on it" },
-        { emoji: "🛠️", text: "Cooking" },
-        { emoji: "🧠", text: "Thinking hard" },
-        { emoji: "🔧", text: "Tinkering" },
-        { emoji: "🚀", text: "On it" },
-        { emoji: "🔥", text: "In the zone" },
-      ]
-      return pairs.some((p) => p.emoji === emoji && p.text === text)
-    },
-    queries: {
-      session: {
-        getValidSession: (db: unknown, token: string) => mockGetValidSession(db, token),
-        getValidSessionWithIdentity: (db: unknown, token: string) => mockGetValidSessionWithIdentity(db, token),
-      },
-      machineToken: {
-        getMachineTokenByToken: (...a: any[]) => mockGetMachineTokenByToken(...a),
-        getLatestTokenForUser: (...a: any[]) => mockGetLatestTokenForUser(...a),
-      },
-      runtime: { getRuntimeIdsByDaemon: (...a: any[]) => mockGetRuntimeIdsByDaemon(...a) },
-      machine: { getMachineByDaemon: (...a: any[]) => mockGetMachineByDaemon(...a) },
-      communityMachine: {
-        hashCredential: (bearer: string) => mockHashCredential(bearer),
-        findCredentialByHash: (...a: any[]) => mockFindCredentialByHash(...a),
-        getMachineByIdForUser: (...a: any[]) => mockGetMachineByIdForUser(...a),
-        upsertMachineByMachineId: (...a: any[]) => mockUpsertMachineByMachineId(...a),
-        touchMachineHeartbeat: (...a: any[]) => mockTouchMachineHeartbeat(...a),
-        markMachineOffline: (...a: any[]) => mockMarkMachineOffline(...a),
-        markMachineOnlineIfOffline: (...a: any[]) => mockMarkMachineOnlineIfOffline(...a),
-        toSummary: (row: any) => mockToSummary(row),
-        isBotOnline: (...a: [unknown, string]) => mockIsBotOnline(...a),
-        reconcileBotActivityFromRunningAgents: (...a: any[]) =>
-          mockReconcileBotActivityFromRunningAgents(...(a as [unknown, string, string[]])),
-      },
-      communityUserProfile: {
-        updateProfile: (...a: any[]) =>
-          mockUpdateProfile(...(a as [unknown, string, { statusEmoji?: string | null; statusText?: string | null }])),
-        getProfile: (...a: any[]) => mockGetProfile(...(a as [unknown, string])),
-      },
-      communityMember: {
-        getCoMemberUserIds: (...a: [unknown, string]) => mockGetCoMemberUserIds(...a),
-        listMembers: (...a: any[]) => mockListMembers(...a),
-      },
-      communityFriendship: {
-        getFriendUserIds: (...a: [unknown, string]) => mockGetFriendUserIds(...a),
-      },
-      communityChannel: {
-        getChannelForMember: (...a: any[]) => mockGetChannelForMember(...a),
-        getChannelType: (...a: any[]) => mockGetChannelType(...a),
-        listChannelMemberUserIds: (...a: any[]) => mockListChannelMemberUserIds(...a),
-        isChannelPrivate: (...a: any[]) => mockIsChannelPrivate(...a),
-        getPrivateChannelAudienceUserIds: (...a: any[]) => mockGetPrivateChannelAudienceUserIds(...a),
-      },
-      communityMembersResolver: {
-        resolveScopeMemberUserIds: (...a: any[]) => mockResolveScopeMemberUserIds(...a),
-      },
-      communityThread: {
-        listThreadParticipantUserIds: (...a: any[]) => mockListThreadParticipantUserIds(...a),
-      },
-      communityDm: {
-        getDM: (...a: any[]) => mockGetDM(...a),
-      },
-      communityBot: {
-        listBotsForMachine: (...a: [unknown, string]) => mockListBotsForMachine(...a),
-        getBotBinding: (...a: [unknown, string]) => mockGetBotBinding(...a),
-        getBotBindingWithOwner: (...a: [unknown, string]) => mockGetBotBindingWithOwner(...a),
-        touchBotRefreshContext: (...a: [unknown, string, string]) => mockTouchBotRefreshContext(...a),
-      },
-      communityBotAuditLog: {
-        insertBotActivityEventAndPrune: (...a: any[]) => mockInsertBotActivityEventAndPrune(...a),
-        insertBotAuditSessionReset: (...a: any[]) => mockInsertBotAuditSessionReset(...a),
-        insertBotAuditNap: (...a: any[]) => mockInsertBotAuditNap(...a),
-        insertBotAuditModelChanged: (...a: any[]) => mockInsertBotAuditModelChanged(...a),
-        insertBotAuditProviderChanged: (...a: any[]) => mockInsertBotAuditProviderChanged(...a),
-      },
-      user: {
-        getUserInternal: (...a: [unknown, string]) => mockGetUserInternal(...a),
-      },
-    },
-  }
-})
-
-// Import after mocks
-import { WebSocketDurableObject } from "./ws-durable"
-
-const mockStubFetch = vi.fn().mockResolvedValue(new (globalThis.Response as any)(JSON.stringify({ sent: 1 })))
-const mockCheckAliveFetch = vi.fn().mockResolvedValue(new (globalThis.Response as any)(JSON.stringify({ alive: true })))
-
-function createDO() {
-  const { ctx, getWebSockets, storage, store } = createMockCtx()
-  const stubGet = vi.fn().mockReturnValue({ fetch: mockStubFetch })
-  const env = {
-    DB: {} as D1Database,
-    WS_DO: {
-      idFromName: vi.fn().mockReturnValue("mock-do-id"),
-      get: stubGet,
-    } as unknown as DurableObjectNamespace,
-    RATE_LIMIT_DO: {} as DurableObjectNamespace,
-  }
-  const durable = new WebSocketDurableObject(ctx, env)
-  return { durable, ctx, getWebSockets, env, stubGet, storage, store }
-}
+import { afterEach, beforeEach, describe, expect, it, vi } from "vitest"
+import { createMockWebSocket } from "./__mocks__/cf"
+import {
+  CFResponse,
+  cleanupHarness,
+  createDO,
+  flushAsyncWork,
+  mockCheckAliveFetch,
+  mockCreateDb,
+  mockFindCredentialByHash,
+  mockGetBotBinding,
+  mockGetBotBindingWithOwner,
+  mockGetChannelForMember,
+  mockGetChannelType,
+  mockGetCoMemberUserIds,
+  mockGetDM,
+  mockGetFriendUserIds,
+  mockGetLatestTokenForUser,
+  mockGetMachineByDaemon,
+  mockGetMachineByIdForUser,
+  mockGetMachineTokenByToken,
+  mockGetPrivateChannelAudienceUserIds,
+  mockGetProfile,
+  mockGetRuntimeIdsByDaemon,
+  mockGetUserInternal,
+  mockGetValidSession,
+  mockGetValidSessionWithIdentity,
+  mockHashCredential,
+  mockInsertBotActivityEventAndPrune,
+  mockInsertBotAuditModelChanged,
+  mockInsertBotAuditNap,
+  mockInsertBotAuditProviderChanged,
+  mockInsertBotAuditSessionReset,
+  mockIsBotOnline,
+  mockIsChannelPrivate,
+  mockListBotsForMachine,
+  mockListChannelMemberUserIds,
+  mockListMembers,
+  mockListThreadParticipantUserIds,
+  mockMarkMachineOffline,
+  mockMarkMachineOnlineIfOffline,
+  mockReconcileBotActivityFromRunningAgents,
+  mockResolveScopeMemberUserIds,
+  mockStubFetch,
+  mockToSummary,
+  mockTouchBotRefreshContext,
+  mockTouchMachineHeartbeat,
+  mockUpdateProfile,
+  mockUpsertMachineByMachineId,
+  resetHarness
+} from "./ws-durable/test-harness"
 
 describe("WebSocketDurableObject", () => {
-  beforeEach(() => {
-    vi.clearAllMocks()
-    // `clearAllMocks` doesn't undo a `mockResolvedValue` set by a prior test —
-    // re-pin these two to their empty default so presence-audience tests
-    // don't leak state into unrelated auth-flow tests.
-    mockGetCoMemberUserIds.mockResolvedValue([])
-    mockGetFriendUserIds.mockResolvedValue([])
-    mockListBotsForMachine.mockResolvedValue([])
-    mockIsBotOnline.mockResolvedValue(false)
-    mockGetUserInternal.mockResolvedValue(null)
-    mockIsChannelPrivate.mockResolvedValue(false)
-    mockGetPrivateChannelAudienceUserIds.mockResolvedValue([])
-    // Non-thread default so typing uses the shared resolver path; the
-    // thread-typing test overrides this.
-    mockGetChannelType.mockResolvedValue("text")
-    mockListThreadParticipantUserIds.mockResolvedValue([])
-    mockListChannelMemberUserIds.mockResolvedValue([])
-    mockGetBotBinding.mockResolvedValue(null)
-    mockUpdateProfile.mockResolvedValue({})
-    mockGetProfile.mockResolvedValue(null)
-    mockReconcileBotActivityFromRunningAgents.mockResolvedValue([])
-    // Daemon-auth binds the token to a machine row for daemonId+workspace;
-    // default to a present row so auth-flow success tests pass. Tests that
-    // exercise a missing machine override this.
-    mockGetMachineByDaemon.mockResolvedValue({ daemonId: "my-daemon", workspaceId: "sp_ws1" })
-  })
+  beforeEach(() => resetHarness())
+  afterEach(() => cleanupHarness())
 
-  describe("fetch — WebSocket upgrade", () => {
-    it("returns 101 for valid WebSocket upgrade", async () => {
-      const { durable } = createDO()
-      const req = new Request("http://internal/?userId=u1", {
-        headers: { Upgrade: "websocket" },
-      })
-
-      const res = await durable.fetch(req)
-
-      expect(res.status).toBe(101)
-      expect((res as unknown as CFResponse).webSocket).toBeDefined()
-    })
-
-    it("returns 426 for non-WebSocket request", async () => {
-      const { durable } = createDO()
-      const req = new Request("http://internal/")
-
-      const res = await durable.fetch(req)
-
-      expect(res.status).toBe(426)
-    })
-
-    it("attaches an unauthenticated ConnectionState on upgrade", async () => {
-      const { durable, ctx } = createDO()
-      const req = new Request("http://internal/?userId=u1", {
-        headers: { Upgrade: "websocket" },
-      })
-
-      await durable.fetch(req)
-
-      const acceptCall = (ctx.acceptWebSocket as ReturnType<typeof vi.fn>).mock.calls[0]
-      const serverWs = acceptCall[0]
-      expect(serverWs.deserializeAttachment()).toEqual({ type: "user", userId: "", authenticated: false })
-    })
-  })
-
-  describe("fetch — broadcast", () => {
-    it("sends message to all authenticated connections", async () => {
-      const { durable, ctx } = createDO()
-
-      // Set up two WebSockets: one authenticated, one not
-      const wsAuth = createMockWebSocket()
-      wsAuth.serializeAttachment({ type: "user", userId: "u1", authenticated: true })
-      const wsUnauth = createMockWebSocket()
-      wsUnauth.serializeAttachment({ type: "user", userId: "", authenticated: false })
-        ; (ctx.getWebSockets as ReturnType<typeof vi.fn>).mockReturnValue([wsAuth, wsUnauth])
-
-      const req = new Request("http://internal/broadcast", {
-        method: "POST",
-        body: JSON.stringify({ type: "runtime.status", daemonId: "d1", workspaceId: "w1", status: "online" }),
-      })
-
-      const res = await durable.fetch(req)
-
-      expect(res.status).toBe(200)
-      expect(await res.json()).toEqual({ sent: 1 })
-      expect(wsAuth.send).toHaveBeenCalledWith(
-        JSON.stringify({ type: "runtime.status", daemonId: "d1", workspaceId: "w1", status: "online" })
-      )
-      expect(wsUnauth.send).not.toHaveBeenCalled()
-    })
-
-    it("returns sent: 0 when no connections exist", async () => {
-      const { durable, ctx } = createDO()
-        ; (ctx.getWebSockets as ReturnType<typeof vi.fn>).mockReturnValue([])
-
-      const req = new Request("http://internal/broadcast", {
-        method: "POST",
-        body: '{"type":"test"}',
-      })
-
-      const res = await durable.fetch(req)
-
-      expect(res.status).toBe(200)
-      expect(await res.json()).toEqual({ sent: 0 })
-    })
-
-    it("skips connections that throw on send (already closed)", async () => {
-      const { durable, ctx } = createDO()
-
-      const wsOpen = createMockWebSocket(WebSocket.OPEN)
-      wsOpen.serializeAttachment({ type: "user", userId: "u1", authenticated: true })
-      const wsClosed = createMockWebSocket(WebSocket.CLOSED)
-      wsClosed.serializeAttachment({ type: "user", userId: "u1", authenticated: true })
-      wsClosed.send.mockImplementation(() => { throw new Error("Connection closed") })
-        ; (ctx.getWebSockets as ReturnType<typeof vi.fn>).mockReturnValue([wsOpen, wsClosed])
-
-      const req = new Request("http://internal/broadcast", {
-        method: "POST",
-        body: '{"type":"test"}',
-      })
-
-      const res = await durable.fetch(req)
-
-      expect(wsOpen.send).toHaveBeenCalled()
-      expect(wsClosed.send).toHaveBeenCalled()
-      expect(await res.json()).toEqual({ sent: 1 })
-    })
-  })
-
-  describe("fetch — /check-user-online (bot-aware, keyed by ?userId=)", () => {
-    // This DO instance is keyed by `user:<id>` (idFromName) but can't read
-    // its own name back off `ctx` on this worker's pinned compatibility_date
-    // — see plans/community-account-debt-fixes.md Fix 3 — so every caller
-    // passes `?userId=` explicitly and the handler branches on it.
-    it("answers via isBotOnline for a bot id, bypassing the live-socket check entirely", async () => {
-      const { durable, ctx } = createDO()
-      mockGetUserInternal.mockResolvedValue({ isBot: true } as any)
-      mockIsBotOnline.mockResolvedValue(true)
-        ; (ctx.getWebSockets as ReturnType<typeof vi.fn>).mockReturnValue([])
-
-      const res = await durable.fetch(new Request("http://internal/check-user-online?userId=bot-1"))
-
-      expect(await res.json()).toEqual({ online: true })
-      expect(mockGetUserInternal).toHaveBeenCalledWith({}, "bot-1")
-      expect(mockIsBotOnline).toHaveBeenCalledWith({}, "bot-1")
-    })
-
-    it("answers false for a bot id with no bound machine or an offline one", async () => {
-      const { durable } = createDO()
-      mockGetUserInternal.mockResolvedValue({ isBot: true } as any)
-      mockIsBotOnline.mockResolvedValue(false)
-
-      const res = await durable.fetch(new Request("http://internal/check-user-online?userId=bot-1"))
-
-      expect(await res.json()).toEqual({ online: false })
-    })
-
-    it("falls back to the live-socket check for a human id (regression: query-param change must not break humans)", async () => {
-      const { durable, ctx } = createDO()
-      mockGetUserInternal.mockResolvedValue({ isBot: false } as any)
-      const wsAuth = createMockWebSocket()
-      wsAuth.serializeAttachment({ type: "user", userId: "human-1", authenticated: true })
-        ; (ctx.getWebSockets as ReturnType<typeof vi.fn>).mockReturnValue([wsAuth])
-
-      const res = await durable.fetch(new Request("http://internal/check-user-online?userId=human-1"))
-
-      expect(await res.json()).toEqual({ online: true })
-      expect(mockIsBotOnline).not.toHaveBeenCalled()
-    })
-
-    it("falls back to the live-socket check when userId is missing entirely", async () => {
-      const { durable, ctx } = createDO()
-      const wsAuth = createMockWebSocket()
-      wsAuth.serializeAttachment({ type: "user", userId: "u1", authenticated: true })
-        ; (ctx.getWebSockets as ReturnType<typeof vi.fn>).mockReturnValue([wsAuth])
-
-      const res = await durable.fetch(new Request("http://internal/check-user-online"))
-
-      expect(await res.json()).toEqual({ online: true })
-      expect(mockGetUserInternal).not.toHaveBeenCalled()
-    })
-
-    it("degrades to { online: false, stale: true } when D1 throws on the bot-online lookup", async () => {
-      // Fail-closed via `readOrStale`: getUserInternal (or isBotOnline)
-      // throwing on retry-exhaust must NOT surface as a 500 — the caller
-      // reads `{ online: false, stale: true }` and moves on.
-      const { durable } = createDO()
-      mockGetUserInternal.mockRejectedValue(new Error("D1 down"))
-
-      const res = await durable.fetch(new Request("http://internal/check-user-online?userId=bot-1"))
-      expect(res.status).toBe(200)
-      expect(await res.json()).toEqual({ online: false, stale: true })
-    })
-  })
-
-  describe("webSocketMessage — auth flow", () => {
-    it("authenticates with valid token and sends auth.ok", async () => {
-      const { durable } = createDO()
-      mockGetValidSessionWithIdentity.mockResolvedValue({ userId: "user-42", name: "Ana", discriminator: "0012" })
-
-      const ws = createMockWebSocket()
-      ws.serializeAttachment({ type: "user", userId: "", authenticated: false })
-
-      await durable.webSocketMessage(ws as any, JSON.stringify({ type: "auth", token: "valid-token" }))
-
-      expect(mockGetValidSessionWithIdentity).toHaveBeenCalledWith({}, "valid-token")
-      expect(ws.send).toHaveBeenCalledWith(JSON.stringify({ type: "auth.ok" }))
-      expect(ws.deserializeAttachment()).toEqual({ type: "user", userId: "user-42", authenticated: true, name: "Ana", discriminator: "0012" })
-    })
-
-    it("closes with 1008 on invalid token", async () => {
-      const { durable } = createDO()
-      mockGetValidSessionWithIdentity.mockResolvedValue(null)
-
-      const ws = createMockWebSocket()
-      ws.serializeAttachment({ type: "user", userId: "", authenticated: false })
-
-      await durable.webSocketMessage(ws as any, JSON.stringify({ type: "auth", token: "bad" }))
-
-      expect(ws.close).toHaveBeenCalledWith(1008, "Unauthorized")
-      expect(ws.send).not.toHaveBeenCalled()
-    })
-
-    it("closes with 1008 when auth message has no token", async () => {
-      const { durable } = createDO()
-
-      const ws = createMockWebSocket()
-      ws.serializeAttachment({ type: "user", userId: "", authenticated: false })
-
-      await durable.webSocketMessage(ws as any, JSON.stringify({ type: "auth" }))
-
-      expect(ws.close).toHaveBeenCalledWith(1008, "Unauthorized")
-      expect(ws.send).not.toHaveBeenCalled()
-      expect(mockGetValidSession).not.toHaveBeenCalled()
-    })
-
-    it("closes with 1008 when auth message has empty string token", async () => {
-      const { durable } = createDO()
-
-      const ws = createMockWebSocket()
-      ws.serializeAttachment({ type: "user", userId: "", authenticated: false })
-
-      await durable.webSocketMessage(ws as any, JSON.stringify({ type: "auth", token: "" }))
-
-      expect(ws.close).toHaveBeenCalledWith(1008, "Unauthorized")
-      expect(ws.send).not.toHaveBeenCalled()
-      expect(mockGetValidSession).not.toHaveBeenCalled()
-    })
-
-    it("closes unauthenticated connection sending non-auth message", async () => {
-      const { durable } = createDO()
-
-      const ws = createMockWebSocket()
-      ws.serializeAttachment({ type: "user", userId: "", authenticated: false })
-
-      await durable.webSocketMessage(ws as any, JSON.stringify({ type: "some-event" }))
-
-      expect(ws.close).toHaveBeenCalledWith(1008, "Not authenticated")
-    })
-
-    it("closes with 1008 when session token is expired (getValidSession returns null)", async () => {
-      const { durable } = createDO()
-      mockGetValidSessionWithIdentity.mockResolvedValue(null)
-
-      const ws = createMockWebSocket()
-      ws.serializeAttachment({ type: "user", userId: "", authenticated: false })
-
-      await durable.webSocketMessage(ws as any, JSON.stringify({ type: "auth", token: "expired-token" }))
-
-      expect(ws.close).toHaveBeenCalledWith(1008, "Unauthorized")
-      expect(ws.send).not.toHaveBeenCalled()
-      expect(ws.deserializeAttachment()).toEqual({ type: "user", userId: "", authenticated: false })
-    })
-
-    it("closes on invalid JSON", async () => {
-      const { durable } = createDO()
-
-      const ws = createMockWebSocket()
-
-      await durable.webSocketMessage(ws as any, "not-json")
-
-      expect(ws.close).toHaveBeenCalledWith(1008, "Invalid JSON")
-    })
-
-    it("ignores binary messages", async () => {
-      const { durable } = createDO()
-
-      const ws = createMockWebSocket()
-
-      await durable.webSocketMessage(ws as any, new ArrayBuffer(4))
-
-      expect(ws.close).not.toHaveBeenCalled()
-      expect(ws.send).not.toHaveBeenCalled()
-    })
-  })
-
-  describe("presence audience — co-members ∪ friends (deduped)", () => {
-    // Presence fan-out must reach friends who share no server, not just
-    // co-members — that's the whole point of a friends list. Exercised
-    // directly against the private helper/methods (bypassing the
-    // fire-and-forget `.catch(() => {})` call sites in the auth flow) so
-    // these assertions aren't racing an un-awaited promise.
-    type PresenceInternals = {
-      getPresenceAudience(userId: string): Promise<string[]>
-      broadcastPresence(userId: string, online: boolean): Promise<void>
-      sendPresenceSnapshot(ws: WebSocket, userId: string): Promise<void>
-    }
-
-    it("getPresenceAudience merges co-members and friends without duplicates", async () => {
-      const { durable } = createDO()
-      mockGetCoMemberUserIds.mockResolvedValue(["member-a", "shared-b"])
-      mockGetFriendUserIds.mockResolvedValue(["shared-b", "friend-c"])
-
-      const audience = await (durable as unknown as PresenceInternals).getPresenceAudience("user-1")
-
-      expect(new Set(audience)).toEqual(new Set(["member-a", "shared-b", "friend-c"]))
-      expect(audience).toHaveLength(3)
-      expect(mockGetCoMemberUserIds).toHaveBeenCalledWith({}, "user-1")
-      expect(mockGetFriendUserIds).toHaveBeenCalledWith({}, "user-1")
-    })
-
-    it("getPresenceAudience returns [] when the user has no co-members and no friends", async () => {
-      const { durable } = createDO()
-      mockGetCoMemberUserIds.mockResolvedValue([])
-      mockGetFriendUserIds.mockResolvedValue([])
-
-      const audience = await (durable as unknown as PresenceInternals).getPresenceAudience("user-1")
-
-      expect(audience).toEqual([])
-    })
-
-    it("broadcastPresence fans out to a friend who shares no server", async () => {
-      const { durable, env } = createDO()
-      mockGetCoMemberUserIds.mockResolvedValue([])
-      mockGetFriendUserIds.mockResolvedValue(["friend-c"])
-      mockStubFetch.mockClear()
-
-      await (durable as unknown as PresenceInternals).broadcastPresence("user-1", true)
-
-      expect(env.WS_DO.idFromName).toHaveBeenCalledWith("user:friend-c")
-      expect(mockStubFetch).toHaveBeenCalledTimes(1)
-      const [req] = mockStubFetch.mock.calls[0] as [Request]
-      expect(req.url).toBe("http://internal/broadcast")
-    })
-
-    it("broadcastPresence no-ops (no fetches) when co-members and friends are both empty", async () => {
-      const { durable } = createDO()
-      mockGetCoMemberUserIds.mockResolvedValue([])
-      mockGetFriendUserIds.mockResolvedValue([])
-      mockStubFetch.mockClear()
-
-      await (durable as unknown as PresenceInternals).broadcastPresence("user-1", true)
-
-      expect(mockStubFetch).not.toHaveBeenCalled()
-    })
-
-    it("sendPresenceSnapshot reports an online friend who shares no server", async () => {
-      const { durable } = createDO()
-      mockGetCoMemberUserIds.mockResolvedValue([])
-      mockGetFriendUserIds.mockResolvedValue(["friend-c"])
-      mockStubFetch.mockResolvedValue(
-        new (globalThis.Response as any)(JSON.stringify({ online: true })),
-      )
-      const ws = createMockWebSocket()
-
-      await (durable as unknown as PresenceInternals).sendPresenceSnapshot(ws as any, "user-1")
-
-      expect(ws.send).toHaveBeenCalledWith(
-        JSON.stringify({ type: "community:presence.update", userId: "friend-c", online: true }),
-      )
-    })
-
-    // Regression (post-Fix-3 hotfix): a fresh bot has no server membership,
-    // so co-members alone can be empty for it even while its bound machine
-    // is genuinely online. The fix lives one layer down, in
-    // `getFriendUserIds` itself (see `community-friendship.test.ts`) — the
-    // owner↔own-bot implicit friendship it already returns for `listFriends`
-    // /`areFriends` now also flows through here, so `getPresenceAudience`
-    // needs no bot-specific branch at all; it just trusts whatever
-    // `getFriendUserIds` (mocked as `mockGetFriendUserIds` above) says.
-  })
-
-  describe("notifyUserDO — bot fan-out on community:machine.status (Fix 3)", () => {
-    // `notifyUserDO` is private; every `community:machine.status` emission in
-    // this file funnels through it (the single choke point the plan calls
-    // for), so exercising it directly here covers all 5 call sites without
-    // duplicating their individual setup.
-    type NotifyInternals = { notifyUserDO(userId: string, payload: unknown): Promise<void> }
-
-    /** Collects `{ url, body }` for every `mockStubFetch` call so far. */
-    async function capturedRequests(): Promise<Array<{ url: string; body: string }>> {
-      return Promise.all(
-        mockStubFetch.mock.calls.map(async ([req]) => ({
-          url: (req as Request).url,
-          body: await (req as Request).clone().text(),
-        }))
-      )
-    }
-
-    it("fans out community:presence.update(online: true) to every bot bound to the machine, beyond the owner notify", async () => {
-      const { durable, env } = createDO()
-      mockListBotsForMachine.mockResolvedValue([
-        { id: "bot-1", name: "Bot One", discriminator: "0001", description: "" },
-        { id: "bot-2", name: "Bot Two", discriminator: "0002", description: "" },
-      ])
-      mockGetCoMemberUserIds.mockResolvedValue(["viewer-1"])
-      mockGetFriendUserIds.mockResolvedValue([])
-      mockStubFetch.mockClear()
-
-      await (durable as unknown as NotifyInternals).notifyUserDO("owner-1", {
-        type: "community:machine.status",
-        machineId: "m1",
-        status: "online",
-        lastSeenAt: "2026-01-01T00:00:00.000Z",
-      })
-
-      expect(mockListBotsForMachine).toHaveBeenCalledWith({}, "m1")
-      expect(env.WS_DO.idFromName).toHaveBeenCalledWith("user:owner-1")
-      expect(env.WS_DO.idFromName).toHaveBeenCalledWith("user:viewer-1")
-
-      const requests = await capturedRequests()
-      // 1 owner notify (raw payload) + 1 broadcastPresence fetch per bot to the shared viewer.
-      expect(requests).toHaveLength(3)
-      expect(requests.filter((r) => r.body.includes('"userId":"bot-1"') && r.body.includes('"online":true'))).toHaveLength(1)
-      expect(requests.filter((r) => r.body.includes('"userId":"bot-2"') && r.body.includes('"online":true'))).toHaveLength(1)
-    })
-
-    it("broadcasts online: false for every bound bot on a machine-offline transition", async () => {
-      const { durable } = createDO()
-      mockListBotsForMachine.mockResolvedValue([
-        { id: "bot-1", name: "Bot One", discriminator: "0001", description: "" },
-      ])
-      mockGetCoMemberUserIds.mockResolvedValue(["viewer-1"])
-      mockGetFriendUserIds.mockResolvedValue([])
-      mockStubFetch.mockClear()
-
-      await (durable as unknown as NotifyInternals).notifyUserDO("owner-1", {
-        type: "community:machine.status",
-        machineId: "m1",
-        status: "offline",
-        lastSeenAt: "2026-01-01T00:00:00.000Z",
-      })
-
-      const requests = await capturedRequests()
-      expect(requests.some((r) => r.body.includes('"userId":"bot-1"') && r.body.includes('"online":false'))).toBe(true)
-    })
-
-    it("a machine bound to zero bots triggers no extra broadcast beyond the owner notify", async () => {
-      const { durable } = createDO()
-      mockListBotsForMachine.mockResolvedValue([])
-      mockStubFetch.mockClear()
-
-      await (durable as unknown as NotifyInternals).notifyUserDO("owner-1", {
-        type: "community:machine.status",
-        machineId: "m1",
-        status: "online",
-        lastSeenAt: "2026-01-01T00:00:00.000Z",
-      })
-
-      expect(mockStubFetch).toHaveBeenCalledTimes(1) // owner notify only
-    })
-
-    it("does not call listBotsForMachine for a payload that isn't a community:machine.status transition", async () => {
-      const { durable } = createDO()
-      mockStubFetch.mockClear()
-
-      await (durable as unknown as NotifyInternals).notifyUserDO("owner-1", {
-        type: "community:machine.updated",
-        machine: { id: "m1" },
-      })
-
-      expect(mockListBotsForMachine).not.toHaveBeenCalled()
-      expect(mockStubFetch).toHaveBeenCalledTimes(1) // owner notify only
-    })
-
-    it("does not throw and skips the bot fan-out on a malformed/non-object payload", async () => {
-      const { durable } = createDO()
-      await expect(
-        (durable as unknown as NotifyInternals).notifyUserDO("owner-1", "not-an-object")
-      ).resolves.toBeUndefined()
-      expect(mockListBotsForMachine).not.toHaveBeenCalled()
-    })
-
-    it("resolves cleanly when the owner-notify stub fetch rejects — the whole method is under try/catch, callers never see the throw", async () => {
-      // Regression guard: the owner-notify `userStub.fetch(...)` used to
-      // sit OUTSIDE the method's try/catch. A stub-fetch throw would
-      // propagate through every caller's `.catch(() => {})` and skip the
-      // bot fan-out entirely, contradicting the comment that claims the
-      // failure would "at least be visible."
-      const { durable } = createDO()
-      mockStubFetch.mockClear()
-      // First call is the owner notify — reject it. Subsequent calls
-      // (bot fan-out per-audience-member) should still fire.
-      mockStubFetch.mockRejectedValueOnce(new Error("stub unreachable"))
-      mockListBotsForMachine.mockResolvedValue([
-        { id: "bot-1", name: "Bot One", discriminator: "0001", description: "" },
-      ])
-      mockGetCoMemberUserIds.mockResolvedValue(["viewer-1"])
-      mockGetFriendUserIds.mockResolvedValue([])
-
-      await expect(
-        (durable as unknown as NotifyInternals).notifyUserDO("owner-1", {
-          type: "community:machine.status",
-          machineId: "m1",
-          status: "online",
-          lastSeenAt: "2026-01-01T00:00:00.000Z",
-        })
-      ).resolves.toBeUndefined()
-
-      // Owner notify was attempted (and failed) — bot fan-out still ran.
-      expect(mockListBotsForMachine).toHaveBeenCalledWith({}, "m1")
-      // Owner notify + at least one broadcastPresence fetch to viewer-1.
-      expect(mockStubFetch.mock.calls.length).toBeGreaterThanOrEqual(2)
-    })
-  })
-
-  describe("webSocketMessage — daemon auth flow", () => {
-    it("rejects daemon with pending token (not yet activated)", async () => {
-      const { durable } = createDO()
-      mockGetMachineTokenByToken.mockResolvedValue({
-        id: "mt_1", userId: "u1", status: "pending", workspaceId: null,
-      })
-
-      const ws = createMockWebSocket()
-      ws.serializeAttachment({ type: "daemon", daemonId: "", userId: "", authenticated: false })
-
-      await durable.webSocketMessage(
-        ws as any,
-        JSON.stringify({ type: "auth", machineToken: "al_pending123", daemonId: "my-daemon" }),
-      )
-
-      expect(ws.close).toHaveBeenCalledWith(1008, "Unauthorized")
-      expect(mockGetRuntimeIdsByDaemon).not.toHaveBeenCalled()
-    })
-
-    it("authenticates daemon with active token (no runtime check in auth)", async () => {
-      const { durable } = createDO()
-      mockGetMachineTokenByToken.mockResolvedValue({
-        id: "mt_1", userId: "u1", status: "active", workspaceId: "sp_ws1",
-      })
-
-      const ws = createMockWebSocket()
-      ws.serializeAttachment({ type: "daemon", daemonId: "", authenticated: false })
-
-      await durable.webSocketMessage(
-        ws as any,
-        JSON.stringify({ type: "auth", machineToken: "al_active123", daemonId: "my-daemon" }),
-      )
-
-      expect(ws.send).toHaveBeenCalledWith(JSON.stringify({ type: "auth.ok" }))
-      // Runtime presence is no longer part of auth — it must not gate the WS
-      // connection (tasks route by workspaceId; runtimes may be mid-register).
-      expect(mockGetRuntimeIdsByDaemon).not.toHaveBeenCalled()
-    })
-
-    it("authenticates daemon with active token even when it has no runtimes yet", async () => {
-      const { durable } = createDO()
-      mockGetMachineTokenByToken.mockResolvedValue({
-        id: "mt_1", userId: "u1", status: "active", workspaceId: "sp_ws1",
-      })
-
-      const ws = createMockWebSocket()
-      ws.serializeAttachment({ type: "daemon", daemonId: "", authenticated: false })
-
-      await durable.webSocketMessage(
-        ws as any,
-        JSON.stringify({ type: "auth", machineToken: "al_noruntimes", daemonId: "my-daemon" }),
-      )
-
-      // No-runtime is a valid state (not an auth failure) — auth.ok, no reject.
-      expect(ws.send).toHaveBeenCalledWith(JSON.stringify({ type: "auth.ok" }))
-      expect(ws.send).not.toHaveBeenCalledWith(JSON.stringify({ type: "error", code: "AUTH_REJECTED" }))
-    })
-
-    it("sends AUTH_REJECTED then closes on definitely-invalid token", async () => {
-      const { durable } = createDO()
-      mockGetMachineTokenByToken.mockResolvedValue(null)
-
-      const ws = createMockWebSocket()
-      ws.serializeAttachment({ type: "daemon", daemonId: "", authenticated: false })
-
-      await durable.webSocketMessage(
-        ws as any,
-        JSON.stringify({ type: "auth", machineToken: "al_dead", daemonId: "my-daemon" }),
-      )
-
-      expect(ws.send).toHaveBeenCalledWith(JSON.stringify({ type: "error", code: "AUTH_REJECTED" }))
-      expect(ws.close).toHaveBeenCalledWith(1008, "Unauthorized")
-    })
-
-    it("closes WITHOUT AUTH_REJECTED on transient D1 failure (daemon must retry)", async () => {
-      const { durable } = createDO()
-      mockGetMachineTokenByToken.mockRejectedValue(new Error("D1 unavailable"))
-
-      const ws = createMockWebSocket()
-      ws.serializeAttachment({ type: "daemon", daemonId: "", authenticated: false })
-
-      await durable.webSocketMessage(
-        ws as any,
-        JSON.stringify({ type: "auth", machineToken: "al_blip", daemonId: "my-daemon" }),
-      )
-
-      expect(ws.send).not.toHaveBeenCalledWith(JSON.stringify({ type: "error", code: "AUTH_REJECTED" }))
-      expect(ws.close).toHaveBeenCalledWith(1011, "Auth temporarily unavailable")
-    })
-
-    it("rejects daemon with unknown token", async () => {
-      const { durable } = createDO()
-      mockGetMachineTokenByToken.mockResolvedValue(null)
-
-      const ws = createMockWebSocket()
-      ws.serializeAttachment({ type: "daemon", daemonId: "", authenticated: false })
-
-      await durable.webSocketMessage(
-        ws as any,
-        JSON.stringify({ type: "auth", machineToken: "al_unknown", daemonId: "my-daemon" }),
-      )
-
-      expect(ws.close).toHaveBeenCalledWith(1008, "Unauthorized")
-    })
-
-    it("rejects daemon with non-al_ prefixed token", async () => {
-      const { durable } = createDO()
-
-      const ws = createMockWebSocket()
-      ws.serializeAttachment({ type: "daemon", daemonId: "", authenticated: false })
-
-      await durable.webSocketMessage(
-        ws as any,
-        JSON.stringify({ type: "auth", machineToken: "bad_prefix", daemonId: "my-daemon" }),
-      )
-
-      expect(ws.close).toHaveBeenCalledWith(1008, "Unauthorized")
-      expect(mockGetMachineTokenByToken).not.toHaveBeenCalled()
-    })
-  })
-
-  describe("webSocketMessage — check_daemon_status (cross-DO)", () => {
-    it("returns runtime.status online when daemon DO reports alive", async () => {
-      const { durable, env } = createDO()
-      mockGetLatestTokenForUser.mockResolvedValue({ hostname: "MyMachine.local" })
-
-      const aliveStub = { fetch: vi.fn().mockResolvedValue(new (globalThis.Response as any)(JSON.stringify({ alive: true }))) }
-        ; (env.WS_DO as any).get = vi.fn().mockReturnValue(aliveStub)
-
-      const ws = createMockWebSocket()
-      ws.serializeAttachment({ type: "user", userId: "user-42", authenticated: true })
-
-      await durable.webSocketMessage(ws as any, JSON.stringify({ type: "check_daemon_status" }))
-
-      expect((env.WS_DO as any).idFromName).toHaveBeenCalledWith("daemon:MyMachine.local")
-      expect(ws.send).toHaveBeenCalledWith(
-        JSON.stringify({ type: "runtime.status", status: "online", daemonId: "MyMachine.local" }),
-      )
-    })
-
-    it("does not respond when daemon DO reports not alive", async () => {
-      const { durable, env } = createDO()
-      mockGetLatestTokenForUser.mockResolvedValue({ hostname: "MyMachine.local" })
-
-      const deadStub = { fetch: vi.fn().mockResolvedValue(new (globalThis.Response as any)(JSON.stringify({ alive: false }))) }
-        ; (env.WS_DO as any).get = vi.fn().mockReturnValue(deadStub)
-
-      const ws = createMockWebSocket()
-      ws.serializeAttachment({ type: "user", userId: "user-42", authenticated: true })
-
-      await durable.webSocketMessage(ws as any, JSON.stringify({ type: "check_daemon_status" }))
-
-      expect(ws.send).not.toHaveBeenCalled()
-    })
-
-    it("does not respond when no token/hostname found", async () => {
-      const { durable } = createDO()
-      mockGetLatestTokenForUser.mockResolvedValue(null)
-
-      const ws = createMockWebSocket()
-      ws.serializeAttachment({ type: "user", userId: "user-42", authenticated: true })
-
-      await durable.webSocketMessage(ws as any, JSON.stringify({ type: "check_daemon_status" }))
-
-      expect(ws.send).not.toHaveBeenCalled()
-    })
-  })
-
-  describe("webSocketError", () => {
-    it("closes with 1011", async () => {
-      const { durable } = createDO()
-
-      const ws = createMockWebSocket()
-
-      await durable.webSocketError(ws as any, new Error("boom"))
-
-      expect(ws.close).toHaveBeenCalledWith(1011, "Internal error")
-    })
-  })
 
   describe("community-machine — session.error overlay + optimistic clear", () => {
     beforeEach(() => {
@@ -1229,6 +156,7 @@ describe("WebSocketDurableObject", () => {
     })
   })
 
+
   describe("community-machine — ready frame wire shape", () => {
     beforeEach(() => {
       mockUpsertMachineByMachineId.mockReset()
@@ -1325,6 +253,7 @@ describe("WebSocketDurableObject", () => {
       expect(mockUpsertMachineByMachineId).not.toHaveBeenCalled()
     })
   })
+
 
   describe("community-machine — agent_activity frame", () => {
     beforeEach(() => {
@@ -1550,6 +479,7 @@ describe("WebSocketDurableObject", () => {
     })
   })
 
+
   describe("community-machine — agent_typing / agent_typing_stop frames", () => {
     beforeEach(() => {
       mockGetBotBindingWithOwner.mockReset()
@@ -1734,6 +664,7 @@ describe("WebSocketDurableObject", () => {
       expect(mockStubFetch.mock.calls.filter((c: any[]) => (c[0] as Request).url.endsWith("/broadcast"))).toHaveLength(1)
     })
   })
+
 
   describe("community-machine — bot_audit_event frame", () => {
     beforeEach(() => {
@@ -1949,6 +880,7 @@ describe("WebSocketDurableObject", () => {
       expect(ws.close as ReturnType<typeof vi.fn>).not.toHaveBeenCalled()
     })
   })
+
 
   describe("community-machine — reset/nap completion re-home (agent_session landing)", () => {
     beforeEach(() => {
@@ -2297,6 +1229,7 @@ describe("WebSocketDurableObject", () => {
     })
   })
 
+
   describe("community-machine — ready frame reconciles bot activity", () => {
     beforeEach(() => {
       mockUpsertMachineByMachineId.mockReset()
@@ -2394,6 +1327,7 @@ describe("WebSocketDurableObject", () => {
       expect(parsed).toEqual([])
     })
   })
+
 
   describe("community-machine — webSocketClose presence lifecycle", () => {
     // These tests cover the "graceful daemon quit → immediate offline" fix.
@@ -2494,6 +1428,7 @@ describe("WebSocketDurableObject", () => {
       expect(ctx.storage.setAlarm).not.toHaveBeenCalled()
     })
   })
+
 
   describe("community-machine — alarm presence + backfill", () => {
     it("live-WS + status=offline row: markMachineOnlineIfOffline flips it back online and broadcasts (post-deploy backfill)", async () => {
@@ -2667,152 +1602,6 @@ describe("WebSocketDurableObject", () => {
       expect(store.has("community-machine-handle")).toBe(true)
       expect(store.has("community-machine-identity")).toBe(true)
       expect(ctx.storage.setAlarm).toHaveBeenCalled()
-    })
-  })
-
-  describe("webSocketMessage — community:typing.start authz (fanOutTyping)", () => {
-    // fanOutTyping runs fire-and-forget (`.catch()`, not awaited) inside
-    // webSocketMessage, so `await durable.webSocketMessage(...)` alone
-    // doesn't guarantee its internal DB-then-broadcast chain has settled.
-    // Flush a macrotask so all pending microtasks (getDM/getChannelForMember
-    // → listMembers → Promise.all(fetch)) drain before asserting.
-    const flush = () => new Promise((resolve) => setTimeout(resolve, 0))
-
-    it("does not fan out or broadcast when sender is not a member of the target channel", async () => {
-      const { durable, env } = createDO()
-      mockGetChannelForMember.mockResolvedValueOnce(null)
-
-      const ws = createMockWebSocket()
-      ws.serializeAttachment({ type: "user", userId: "attacker", authenticated: true })
-
-      await durable.webSocketMessage(
-        ws as any,
-        JSON.stringify({ type: "community:typing.start", channelId: "chan-private" }),
-      )
-      await flush()
-
-      expect(mockGetChannelForMember).toHaveBeenCalledWith(expect.anything(), "chan-private", "attacker")
-      expect(mockListMembers).not.toHaveBeenCalled()
-      expect((env.WS_DO as any).get).not.toHaveBeenCalled()
-      expect(mockStubFetch).not.toHaveBeenCalled()
-    })
-
-    it("fans out to other server members when sender IS a channel member", async () => {
-      const { durable, env } = createDO()
-      mockGetChannelForMember.mockResolvedValueOnce({ id: "chan-1", serverId: "server-1" })
-      // Recipient resolution now goes through the shared member resolver — a
-      // public channel resolves to every server member (sender included).
-      mockResolveScopeMemberUserIds.mockResolvedValueOnce(["member-1", "member-2", "sender-1"])
-
-      const ws = createMockWebSocket()
-      ws.serializeAttachment({ type: "user", userId: "sender-1", authenticated: true })
-
-      await durable.webSocketMessage(
-        ws as any,
-        JSON.stringify({ type: "community:typing.start", channelId: "chan-1" }),
-      )
-      await flush()
-
-      expect(mockGetChannelForMember).toHaveBeenCalledWith(expect.anything(), "chan-1", "sender-1")
-      expect(mockResolveScopeMemberUserIds).toHaveBeenCalledWith(expect.anything(), {
-        scope: "channel",
-        scopeId: "chan-1",
-      })
-      // Sender is excluded from recipients — only the other 2 members get a broadcast POST.
-      expect((env.WS_DO as any).idFromName).toHaveBeenCalledWith("user:member-1")
-      expect((env.WS_DO as any).idFromName).toHaveBeenCalledWith("user:member-2")
-      expect((env.WS_DO as any).idFromName).not.toHaveBeenCalledWith("user:sender-1")
-      expect(mockStubFetch).toHaveBeenCalledTimes(2)
-    })
-
-    it("private channel: fans out to the channel audience, not all server members", async () => {
-      const { durable, env } = createDO()
-      mockGetChannelForMember.mockResolvedValueOnce({ id: "chan-p", serverId: "server-1" })
-      // The resolver applies the public/private split internally — a private
-      // channel resolves to its audience (creator + added member + admin).
-      mockResolveScopeMemberUserIds.mockResolvedValueOnce(["sender-1", "member-1"])
-
-      const ws = createMockWebSocket()
-      ws.serializeAttachment({ type: "user", userId: "sender-1", authenticated: true })
-
-      await durable.webSocketMessage(
-        ws as any,
-        JSON.stringify({ type: "community:typing.start", channelId: "chan-p" }),
-      )
-      await flush()
-
-      expect(mockResolveScopeMemberUserIds).toHaveBeenCalledWith(expect.anything(), {
-        scope: "channel",
-        scopeId: "chan-p",
-      })
-      expect(mockListMembers).not.toHaveBeenCalled()
-      // Only the non-sender audience member gets a broadcast POST.
-      expect((env.WS_DO as any).idFromName).toHaveBeenCalledWith("user:member-1")
-      expect((env.WS_DO as any).idFromName).not.toHaveBeenCalledWith("user:sender-1")
-      expect(mockStubFetch).toHaveBeenCalledTimes(1)
-    })
-
-    it("thread: typing fans out to PARTICIPANTS, not the channel audience", async () => {
-      const { durable, env } = createDO()
-      mockGetChannelForMember.mockResolvedValueOnce({ id: "t-1", serverId: "server-1" })
-      mockGetChannelType.mockResolvedValueOnce("thread")
-      mockListThreadParticipantUserIds.mockResolvedValueOnce(["sender-1", "part-1"])
-
-      const ws = createMockWebSocket()
-      ws.serializeAttachment({ type: "user", userId: "sender-1", authenticated: true })
-
-      await durable.webSocketMessage(
-        ws as any,
-        JSON.stringify({ type: "community:typing.start", channelId: "t-1" }),
-      )
-      await flush()
-
-      expect(mockListThreadParticipantUserIds).toHaveBeenCalledWith(expect.anything(), "t-1")
-      // Thread typing must NOT fall back to the channel-audience resolver.
-      expect(mockResolveScopeMemberUserIds).not.toHaveBeenCalled()
-      expect((env.WS_DO as any).idFromName).toHaveBeenCalledWith("user:part-1")
-      expect((env.WS_DO as any).idFromName).not.toHaveBeenCalledWith("user:sender-1")
-      expect(mockStubFetch).toHaveBeenCalledTimes(1)
-    })
-
-    it("does not fan out when sender is not a participant of the target DM", async () => {
-      const { durable, env } = createDO()
-      // A DM is a type=dm channel now; a non-member sender can't see it, so
-      // getChannelForMember returns null and fanOutTyping bails.
-      mockGetChannelForMember.mockResolvedValueOnce(null)
-
-      const ws = createMockWebSocket()
-      ws.serializeAttachment({ type: "user", userId: "attacker", authenticated: true })
-
-      await durable.webSocketMessage(
-        ws as any,
-        JSON.stringify({ type: "community:typing.start", channelId: "dm-1" }),
-      )
-      await flush()
-
-      expect(mockGetChannelForMember).toHaveBeenCalledWith(expect.anything(), "dm-1", "attacker")
-      expect((env.WS_DO as any).get).not.toHaveBeenCalled()
-      expect(mockStubFetch).not.toHaveBeenCalled()
-    })
-
-    it("fans out to the other participant when sender IS a DM participant", async () => {
-      const { durable, env } = createDO()
-      mockGetChannelForMember.mockResolvedValueOnce({ id: "dm-1", serverId: null })
-      mockGetChannelType.mockResolvedValueOnce("dm")
-      mockListChannelMemberUserIds.mockResolvedValueOnce(["alice", "bob"])
-
-      const ws = createMockWebSocket()
-      ws.serializeAttachment({ type: "user", userId: "alice", authenticated: true })
-
-      await durable.webSocketMessage(
-        ws as any,
-        JSON.stringify({ type: "community:typing.start", channelId: "dm-1" }),
-      )
-      await flush()
-
-      expect((env.WS_DO as any).idFromName).toHaveBeenCalledWith("user:bob")
-      expect((env.WS_DO as any).idFromName).not.toHaveBeenCalledWith("user:alice")
-      expect(mockStubFetch).toHaveBeenCalledTimes(1)
     })
   })
 })
